rename add_graphene to add_graphql
