+++ cookiecutter.json
@@ -18,7 +18,7 @@
     , "add_sentry": "y"
     , "add_django_auth_wall": "y"
     , "add_celery": "n"
-    , "add_graphene": "n"
+    , "add_graphql": "n"
     , "add_pre_commit": "y"
     , "pagination": ["LimitOffsetPagination", "CursorPagination"]
 }
