+++ {{cookiecutter.github_repository}}/pyproject.toml
@@ -34,7 +34,7 @@ djangorestframework = "^3.13"
 drf-yasg = "^1.20"
 
 # Graph APIs
-{% if cookiecutter.add_graphene == "y" -%}
+{% if cookiecutter.add_graphql == "y" -%}
 graphene-django = "^2.15.0"
 django-graphql-jwt = "^0.3.4"
 {%- endif %}
