+++ {{cookiecutter.github_repository}}/{{cookiecutter.main_module}}/users/auth/utils.py
@@ -3,10 +3,6 @@
 from uuid import UUID
 
 # Third Party Stuff
-{%- if cookiecutter.add_graphene == "y" %}
-from django.contrib.auth import get_user_model
-{%- endif %}
-
 from django.utils.encoding import force_bytes, force_str
 from django.utils.http import urlsafe_base64_decode, urlsafe_base64_encode
 
@@ -25,13 +21,3 @@ def decode_uuid_from_base64(uuid_value: str):
         return UUID(force_str(urlsafe_base64_decode(uuid_value)))
     except (ValueError, OverflowError, TypeError):
         return None
-
-
-{%- if cookiecutter.add_graphene == "y" %}
-def get_user_by_id(user_id):
-    user_model = get_user_model()
-    try:
-        return user_model.objects.get(id=user_id)
-    except user_model.DoesNotExist:
-        return None
-{%- endif %}
